Made changes to the PR such that it takes relative path
