+++ Maven/metadata_summary.json
@@ -1,0 +1,20 @@
+[
+    {
+        "creation_time": "2024-11-20T15:23:44.2493537Z",
+        "last_modified_time": "2024-11-20T15:23:44.2493537Z",
+        "file_name": "saq_bluesky_bgd_20211001_20230430_inv_tbe.csv",
+        "file_size": 13522
+    },
+    {
+        "creation_time": "2024-11-20T15:23:44.2493537Z",
+        "last_modified_time": "2024-11-20T15:23:44.2493537Z",
+        "file_name": "saq_bluesky_dku_20210715_20230131_inv_tbe.csv",
+        "file_size": 13227
+    },
+    {
+        "creation_time": "2024-11-20T15:23:44.2493537Z",
+        "last_modified_time": "2024-11-20T15:23:44.2493537Z",
+        "file_name": "saq_bluesky_npl_20220830_20230404_inv_tbe.csv",
+        "file_size": 17720
+    }
+]
