--- python_tbe/Maven/python_metadata_summary.json
@@ -1,19 +1,0 @@
-{
-    "tbe_map_inventory.py": {},
-    "ebs_read_tbe.py": {},
-    "dku_postprocess.py": {
-        "Plot temporal variations by day, week, month": "overlay sites",
-        "Matching data": "do only times which have valid data for all sites"
-    },
-    "dku_correlogram.py": {},
-    "dku_leaflet_map.py": {
-        "timemaxsct = pd.to_datetime('2023-02-15 00": "00')",
-        "timeminsct = pd.to_datetime('2022-12-15 00": "00')"
-    },
-    "bdf_utils.py": {},
-    "dku_boxplot.py": {
-        "# Input directory and file name": "indir, flroot",
-        "# Sites to select": "either specify in asites, or use unique to select all"
-    },
-    "dku_bluesky_analysis.py": {}
-}
